requirements.txt
# Selenium & Web Driver
selenium==4.15.0
webdriver-manager==4.0.1

# Test Framework & Plugins
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
pytest-timeout==2.2.0

# Data Handling
openpyxl==3.1.2
pandas==2.2.0
lxml==4.9.3
pyarrow==16.1.0
python-dotenv==1.0.0
colorlog==6.8.0
urllib3==2.1.0
# Utilities
configparser==6.0.0
Faker==19.12.0
requests==2.31.0
pillow==10.1.0
pyautogui==0.9.54